requirements.txt
chromadb==0.5.23
openai==1.82.0
tiktoken==0.9.0
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
pdfplumber==0.11.6
pandas==2.2.3
openpyxl==3.1.5
